[CLEANUP-3] Fix security_config.py imports and formatting - Issue #153
